code refactor

+++ students-database/Table/Table.cpp
@@ -23,46 +23,44 @@ void Table::formatColumn(string str, int maxLen, int tableWidth, bool header) {
     cout << " |";
 }
 
-void Table::drawLine(List<string> columns, List<int> params, int tableWidth) {
+void Table::drawLine(List<string > columns, List<int> params, int tableWidth) {
     for (int i = 0; i < columns.size(); i++) {
         formatColumn(columns[i], params[i], tableWidth);
     }
 }
 
-int Table:: getMaxLen(List<string> names) {
+int Table::getMaxLen(List<string > names) {
     int max = 0;
-    for (string str : names)
-    {
-        int len = str.size();
-        max = len > max ? len : max;
+    for (string str: names) {
+        size_t len = str.size();
+        max = len > max ? (int)len : max;
     }
     return max + 4;
 }
 
-void Table::DrawTable(List<List<string>> lines, List<string> columns, string title) {
+void Table::DrawTable(List<List<string>> lines, List<string > columns, string title) {
 
     List<int> params;
     for (int i = 0; i < columns.size(); i++) {
-        List<string> columnValues = { columns[i] };
+        List<string > columnValues = {columns[i]};
         for (int j = 0; j < lines.size(); j++) {
             columnValues.add(lines[j][i]);
         }
         params.add(getMaxLen(columnValues));
     }
 
-    //int tableWidth = accumulate(params.begin(), params.end(), 0);
     int tableWidth = params.sum();
     HWND hWindowConsole = GetConsoleWindow();
     RECT r;
     GetWindowRect(hWindowConsole, &r);
-   
+
     if (r.right - r.left < tableWidth * 9) {
         MoveWindow(hWindowConsole, r.left, r.top, tableWidth * 9, 700, TRUE);
     }
 
     drawBorder(tableWidth);
     if (title.size()) {
-        formatColumn(title, title.size(), tableWidth, true);
+        formatColumn(title, (int)title.size(), tableWidth, true);
         drawBorder(tableWidth);
     }
